requirements.txt
# LangGraph + LangChain
langgraph>=0.2.0
langgraph-checkpoint-sqlite>=1.0.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-ollama>=0.2.0

# Data Models
pydantic>=2.0

# Vector Database
pymilvus>=2.4.0

# Server — Socket.IO + aiohttp
python-socketio[aiohttp]>=5.0
aiohttp>=3.0

# gRPC
grpcio>=1.60
grpcio-tools>=1.60

# Utilities
aiofiles>=24.0
aiosqlite>=0.19.0
python-dotenv>=1.0
requests>=2.31
pyright>=1.1.0

# Markdown rendering (backend)
markdown>=3.7
Pygments>=2.19

# Web Search
duckduckgo_search>=5.0

# CLI UI
rich>=13.9.4
prompt_toolkit>=3.0.48
